requirements.txt
face-recognition
opencv-python
numpy
jupyter
pillow
